+++ src/main/java/org/example/pmas/repository/Interfaces/CrudInterface.java
@@ -1,4 +1,4 @@
-package org.example.pmas.repository;
+package org.example.pmas.repository.Interfaces;
 
 import java.util.List;
