+++ src/main/java/org/example/pmas/repository/Interfaces/ProjectRepositoryInterface.java
@@ -1,0 +1,5 @@
+package org.example.pmas.repository.Interfaces;
+
+public interface ProjectRepositoryInterface extends CrudInterface {
+
+}
