+++ src/main/java/org/example/pmas/repository/Interfaces/SubProjectRepositoryInterface.java
@@ -1,0 +1,4 @@
+package org.example.pmas.repository.Interfaces;
+
+public interface SubProjectRepositoryInterface extends CrudInterface {
+}
