+++ src/main/java/org/example/pmas/repository/Interfaces/UserRepositoryInterface.java
@@ -1,0 +1,4 @@
+package org.example.pmas.repository.Interfaces;
+
+public interface UserRepositoryInterface extends CrudInterface {
+}
